On branch master  Your branch is ahead of 'origin/master' by 1 commit.    (use "git push" to publish your local commits)
 Changes to be committed:
	deleted:    1.png
	deleted:    2.png
	deleted:    3.png
	renamed:    qx3/Battlefield.java -> Battlefield.java
	renamed:    qx3/CSFormation.java -> CSFormation.java
	new file:   Calabash.java
	renamed:    qx3/Creature.java -> Creature.java
	renamed:    qx3/FSFormation.java -> FSFormation.java
	renamed:    qx3/FYFormation.java -> FYFormation.java
	renamed:    qx3/Grandpa.java -> Grandpa.java
	renamed:    qx3/HGFormation.java -> HGFormation.java
	renamed:    qx3/HYFormation.java -> HYFormation.java
	renamed:    qx3/Main.java -> Main.java
	renamed:    qx3/Monster.java -> Monster.java
	renamed:    qx3/MonsterFormation.java -> MonsterFormation.java
	modified:   README.md
	renamed:    qx3/Scorpion.java -> Scorpion.java
	renamed:    qx3/Snake.java -> Snake.java
	renamed:    qx3/Underling.java -> Underling.java
	renamed:    qx3/Warrior.java -> Warrior.java
	renamed:    qx3/WarriorFormation.java -> WarriorFormation.java
	renamed:    qx3/YLFormation.java -> YLFormation.java
	renamed:    qx3/YXFormation.java -> YXFormation.java
	renamed:    qx3/YYFormation.java -> YYFormation.java
	deleted:    picture/0.png
	deleted:    picture/1.png
	deleted:    picture/10.png
	deleted:    picture/11.png
	deleted:    picture/2.png
	deleted:    picture/3.png
	deleted:    picture/4.png
	deleted:    picture/5.png
	deleted:    picture/6.png
	deleted:    picture/7.png
	deleted:    picture/8.png
	deleted:    picture/9.png
	deleted:    qx3/Calabash.java
	deleted:    sample/Controller.java
	deleted:    sample/Main.java
	deleted:    sample/sample.fxml
