+++ 20180925/钱鑫-161220099/Creature.java
@@ -1,16 +1,13 @@
-package qx3;
+package qx2;
 
 public abstract class Creature {
     protected String name;
-    protected String url;
     public void setName(String name)
     {
         this.name = name;
     }
-    public void setUrl(String url){this.url=url;}
     public String getName()
     {
         return name;
     }
-    public String getUrl(){return url;}
 }
