+++ 20180925/钱鑫-161220099/Grandpa.java
@@ -1,9 +1,8 @@
-package qx3;
+package qx2;
 
 public class Grandpa extends Creature {
     public Grandpa()
     {
         setName("爷爷");
-        setUrl("/picture/0.png");
     }
 }
