+++ 20180925/钱鑫-161220099/HYFormation.java
@@ -1,21 +1,22 @@
-package qx3;
+package qx2;
 
 import java.util.List;
 
-public class YXFormation implements MonsterFormation {
+public class HYFormation implements MonsterFormation {
     @Override
     public void formation(Monster monster,Battlefield battlefield)
     {
         int n=battlefield.getN();
         Snake snake=monster.getSnake();
         Scorpion scorpion=monster.getScorpion();
-        monster.init(6);
+        monster.init(10);
         List<Underling> underlings=monster.getUnderlings();
         battlefield.setPosition(n-1,n-1,snake);
-        battlefield.setPosition((n-1)/2+3,n-1,scorpion);
-        for(int i=1;i<=6;i++)
+        battlefield.setPosition((n-1)/2,n-1,scorpion);
+        for(int i=1;i<=5;i++)
         {
-            battlefield.setPosition((n-1)/2+3-i,n-1-i,underlings.get(i-1));
+            battlefield.setPosition((n-1)/2-i,n-1-i,underlings.get(i-1));
+            battlefield.setPosition((n-1)/2+i,n-1-i,underlings.get(i-1));
         }
     }
 }
