+++ 20180925/钱鑫-161220099/MonsterFormation.java
@@ -1,4 +1,4 @@
-package qx3;
+package qx2;
 
 public interface MonsterFormation {
     public void formation(Monster monster,Battlefield battlefield);
