+++ 20180925/钱鑫-161220099/Scorpion.java
@@ -1,9 +1,8 @@
-package qx3;
+package qx2;
 
 public class Scorpion extends Creature {
     public Scorpion()
     {
         setName("蝎精");
-        setUrl("/picture/8.png");
     }
 }
