+++ 20180925/钱鑫-161220099/Snake.java
@@ -1,9 +1,8 @@
-package qx3;
+package qx2;
 
 public class Snake extends Creature {
     public Snake()
     {
         setName("蛇精");
-        setUrl("/picture/9.png");
     }
 }
