+++ 20180925/钱鑫-161220099/WarriorFormation.java
@@ -1,4 +1,4 @@
-package qx3;
+package qx2;
 
 import java.util.List;
